Proper mesh generation

+++ Minecraft/Minecraft/Client/World/Chunk.h
@@ -11,7 +11,7 @@ namespace Minecraft::Terrain {
 
 		void generateData(int relX, int relZ, int relY);
 		void generateMesh();
-		void tryAddFaceToMesh(const float blockFace[12], std::array<float, 8> texCoords, const mc::Vector3i& blockPosition, const mc::Vector3i& blockFacing, int type);
+		void tryAddFaceToMesh(const short blockFace[12], std::array<float, 8> texCoords, const mc::Vector3i& blockPosition, const mc::Vector3i& blockFacing, int type);
 
 		void Update();
 		void Draw();
